requirements.txt
cycler==0.11.0
fonttools==4.34.4
graphviz==0.20
kiwisolver==1.4.3
matplotlib==3.5.2
neat-python==0.92
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pygame==2.1.2
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
six==1.16.0
